% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glsMs.R
\name{glsMs}
\alias{glsMs}
\title{GLS model selection for time series}
\usage{
glsMs(formula, data, diagnostic = F, list_models = F, ...)
}
\arguments{
\item{formula}{An object of class \code{formula} describing the model to be fitted. Should be
in the form \code{series ~ time}.}

\item{data}{Input data.frame to be analyzed.}

\item{diagnostic}{\code{logical}. Returns best fitting model from selection process if \code{TRUE}.}

\item{list_models}{\code{logical}. Returns coefficients and AICc scores of all fitted models.}

\item{...}{Other arguments may be passed to the gls.}
}
\value{
gls model object
}
\description{
GLS model selection for time series
}
\examples{
#Generate series

m <- 0.1
x <- 1:30

data <- data.frame(x = x,
                  y = m*x + rnorm(30, sd = 0.35))
glsMs(formula = y ~ x, data = data, diagnostic = T)
}
